feat : update useModal of postUserSignUp hook
