+++ apps/client/src/app/auth/_state/server/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./usePostUserSignUp";
