+++ apps/client/src/app/auth/_state/server/index.ts
@@ -1,1 +1,1 @@
-export * from "./postUserSignUp";
+export * from "./hooks";
